Shell script to study the ue of Diff utility
Created a text file “a.txt” . Copy this file to another file “b.txt”. Modified a few lines here and there in b.txt. Use diff utility to find the difference between both files. Redirect the output of diff to a file named a.patch. Applied the patch file on a.txt.